remaning the new use case
